Renaming again to get changes in folders published on Netlify
